68 gui message bus rework

+++ src/stock_watch/data_scraper/scraper_process.py
@@ -1,6 +1,7 @@
 from .scrapers.scraper import Scraper
 import multiprocessing
 import logging
+from multiprocessing.connection import Connection
 
 
 class ScraperProcess(object):
@@ -10,28 +11,38 @@ class ScraperProcess(object):
         process.
         :param scraper:
         """
+        self.parent_conn = None
+        self.child_conn = None
         self._scraper = scraper
         self._running = False
         self._process = None
 
-    def start(self):
+    def start(self) -> Connection:
         """
         Start the scraper in a separate process
         :return:
         """
         self._running = True
-        self._process = multiprocessing.Process(target=self._scraper.start)
+        self.parent_conn, self.child_conn = multiprocessing.Pipe(duplex=True)
+        self._process = multiprocessing.Process(target=self._scraper.start, args=(self.child_conn,))
         self._process.start()
         logging.info(f"Started {self._scraper.__class__.__name__}")
+
+        return self.parent_conn
+
 
     def stop(self):
         """
         Stop the scraper process
         :return:
         """
-        self._running = False
-        self._process.terminate()
-        logging.info(f"Stopped {self._scraper.__class__.__name__}")
+        if self._running:
+            self._running = False
+            self._process.terminate()
+            self._process.join()
+            logging.info(f"Stopped {self._scraper.__class__.__name__}")
+        else:
+            logging.warning(f"{self._scraper.__class__.__name__} is not running")
 
     @property
     def running(self):
